After update from Bluehost.

+++ wp-content/plugins/wordfence/lib/wfConfig.php
@@ -34,6 +34,7 @@ class wfConfig {
 				"scansEnabled_plugins" => false,
 				"scansEnabled_malware" => false,
 				"scansEnabled_fileContents" => false,
+				"scansEnabled_database" => false,
 				"scansEnabled_posts" => false,
 				"scansEnabled_comments" => false,
 				"scansEnabled_passwds" => false,
@@ -69,6 +70,7 @@ class wfConfig {
 				"debugOn" => false,
 				'email_summary_enabled' => true,
 				'email_summary_dashboard_widget_enabled' => true,
+				'ssl_verify' => true,
 			),
 			"otherParams" => array(
 				'securityLevel' => '0',
@@ -94,6 +96,7 @@ class wfConfig {
 				'blockedTime' => "300",
 				'email_summary_interval' => 'biweekly',
 				'email_summary_excluded_directories' => 'wp-content/cache,wp-content/wfcache,wp-content/plugins/wordfence/tmp',
+				'allowed404s' => "/favicon.ico\n/apple-touch-icon*.png\n/*@2x.png",
 			)
 		),
 		array( //level 1
@@ -121,6 +124,7 @@ class wfConfig {
 				"scansEnabled_plugins" => false,
 				"scansEnabled_malware" => true,
 				"scansEnabled_fileContents" => true,
+				"scansEnabled_database" => true,
 				"scansEnabled_posts" => true,
 				"scansEnabled_comments" => true,
 				"scansEnabled_passwds" => true,
@@ -156,6 +160,7 @@ class wfConfig {
 				"debugOn" => false,
 				'email_summary_enabled' => true,
 				'email_summary_dashboard_widget_enabled' => true,
+				'ssl_verify' => true,
 			),
 			"otherParams" => array(
 				'securityLevel' => '1',
@@ -181,6 +186,7 @@ class wfConfig {
 				'blockedTime' => "300",
 				'email_summary_interval' => 'biweekly',
 				'email_summary_excluded_directories' => 'wp-content/cache,wp-content/wfcache,wp-content/plugins/wordfence/tmp',
+				'allowed404s' => "/favicon.ico\n/apple-touch-icon*.png\n/*@2x.png",
 			)
 		),
 		array( //level 2
@@ -208,6 +214,7 @@ class wfConfig {
 				"scansEnabled_plugins" => false,
 				"scansEnabled_malware" => true,
 				"scansEnabled_fileContents" => true,
+				"scansEnabled_database" => true,
 				"scansEnabled_posts" => true,
 				"scansEnabled_comments" => true,
 				"scansEnabled_passwds" => true,
@@ -243,6 +250,7 @@ class wfConfig {
 				"debugOn" => false,
 				'email_summary_enabled' => true,
 				'email_summary_dashboard_widget_enabled' => true,
+				'ssl_verify' => true,
 			),
 			"otherParams" => array(
 				'securityLevel' => '2',
@@ -268,6 +276,7 @@ class wfConfig {
 				'blockedTime' => "300",
 				'email_summary_interval' => 'biweekly',
 				'email_summary_excluded_directories' => 'wp-content/cache,wp-content/wfcache,wp-content/plugins/wordfence/tmp',
+				'allowed404s' => "/favicon.ico\n/apple-touch-icon*.png\n/*@2x.png",
 			)
 		),
 		array( //level 3
@@ -295,6 +304,7 @@ class wfConfig {
 				"scansEnabled_plugins" => false,
 				"scansEnabled_malware" => true,
 				"scansEnabled_fileContents" => true,
+				"scansEnabled_database" => true,
 				"scansEnabled_posts" => true,
 				"scansEnabled_comments" => true,
 				"scansEnabled_passwds" => true,
@@ -330,6 +340,7 @@ class wfConfig {
 				"debugOn" => false,
 				'email_summary_enabled' => true,
 				'email_summary_dashboard_widget_enabled' => true,
+				'ssl_verify' => true,
 			),
 			"otherParams" => array(
 				'securityLevel' => '3',
@@ -355,6 +366,7 @@ class wfConfig {
 				'blockedTime' => "1800",
 				'email_summary_interval' => 'biweekly',
 				'email_summary_excluded_directories' => 'wp-content/cache,wp-content/wfcache,wp-content/plugins/wordfence/tmp',
+				'allowed404s' => "/favicon.ico\n/apple-touch-icon*.png\n/*@2x.png",
 			)
 		),
 		array( //level 4
@@ -382,6 +394,7 @@ class wfConfig {
 				"scansEnabled_plugins" => false,
 				"scansEnabled_malware" => true,
 				"scansEnabled_fileContents" => true,
+				"scansEnabled_database" => true,
 				"scansEnabled_posts" => true,
 				"scansEnabled_comments" => true,
 				"scansEnabled_passwds" => true,
@@ -417,6 +430,7 @@ class wfConfig {
 				"debugOn" => false,
 				'email_summary_enabled' => true,
 				'email_summary_dashboard_widget_enabled' => true,
+				'ssl_verify' => true,
 			),
 			"otherParams" => array(
 				'securityLevel' => '4',
@@ -442,6 +456,7 @@ class wfConfig {
 				'blockedTime' => "7200",
 				'email_summary_interval' => 'biweekly',
 				'email_summary_excluded_directories' => 'wp-content/cache,wp-content/wfcache,wp-content/plugins/wordfence/tmp',
+				'allowed404s' => "/favicon.ico\n/apple-touch-icon*.png\n/*@2x.png",
 			)
 		)
 	);
@@ -492,7 +507,7 @@ class wfConfig {
 		}
 		foreach(self::$securityLevels[2]['otherParams'] as $key => $val){
 			if(isset($_POST[$key])){
-				$ret[$key] = $_POST[$key];
+				$ret[$key] = stripslashes($_POST[$key]);
 			} else {
 				error_log("Missing options param \"$key\" when parsing parameters.");
 			}
@@ -515,7 +530,7 @@ class wfConfig {
 		self::$cache = array();
 	}
 	public static function getHTML($key){
-		return wp_kses(self::get($key), array());
+		return esc_html(self::get($key));
 	}
 	public static function inc($key){
 		$val = self::get($key, false);
@@ -569,16 +584,13 @@ class wfConfig {
 			}
 		}
 
-		if(! isset(self::$cache[$key])){ 
+		if(!array_key_exists($key, self::$cache)){ 
 			$val = self::loadFromDiskCache($key);
 			//$val = self::getDB()->querySingle("select val from " . self::table() . " where name='%s'", $key);
-			if(isset($val)){
-				self::$cache[$key] = $val;
-			} else {
-				self::$cache[$key] = $default;
-			}
+			self::$cache[$key] = $val;
 		}
-		return self::$cache[$key];
+		$val = self::$cache[$key];
+		return $val !== null ? $val : $default;
 	}
 	public static function loadFromDiskCache($key){
 		if(! self::$diskCacheDisabled){
@@ -789,8 +801,11 @@ class wfConfig {
 		return true;
 	}
 	public static function enableAutoUpdate(){
-		wfConfig::set('autoUpdate', '1');	
-		wp_schedule_event(time(), 'daily', 'wordfence_daily_autoUpdate');
+		wfConfig::set('autoUpdate', '1');
+		wp_clear_scheduled_hook('wordfence_daily_autoUpdate');
+		if (is_main_site()) {
+			wp_schedule_event(time(), 'daily', 'wordfence_daily_autoUpdate');
+		}
 	}
 	public static function disableAutoUpdate(){
 		wfConfig::set('autoUpdate', '0');	
@@ -857,11 +872,12 @@ Options -ExecCGI
 		$upload_dir = wp_upload_dir();
 		return $upload_dir['basedir'] . '/.htaccess';
 	}
-	
+
 	/**
 	 * Add/Merge .htaccess file in the uploads directory to prevent code execution.
 	 *
 	 * @return bool
+	 * @throws wfConfigException
 	 */
 	public static function disableCodeExecutionForUploads() {
 		$uploads_htaccess_file_path = self::_uploadsHtaccessFilePath();
@@ -880,27 +896,32 @@ Options -ExecCGI
 		}
 		return true;
 	}
-	
+
 	/**
 	 * Remove script execution protections for our the .htaccess file in the uploads directory.
 	 *
 	 * @return bool
+	 * @throws wfConfigException
 	 */
 	public static function removeCodeExecutionProtectionForUploads() {
 		$uploads_htaccess_file_path = self::_uploadsHtaccessFilePath();
 		if (file_exists($uploads_htaccess_file_path)) {
 			$htaccess_contents = file_get_contents($uploads_htaccess_file_path);
-			$htaccess_contents = str_replace(self::$_disable_scripts_htaccess, '', $htaccess_contents);
-			
-			$error_message = "Unable to remove code execution protections applied to the .htaccess file in the uploads directory.  Please check your permissions on that file.";
-			if (strlen(trim($htaccess_contents)) === 0) {
-				// empty file, remove it
-				if (!@unlink($uploads_htaccess_file_path)) {
+
+			// Check that it is in the file
+			if (strpos($htaccess_contents, self::$_disable_scripts_htaccess) !== false) {
+				$htaccess_contents = str_replace(self::$_disable_scripts_htaccess, '', $htaccess_contents);
+
+				$error_message = "Unable to remove code execution protections applied to the .htaccess file in the uploads directory.  Please check your permissions on that file.";
+				if (strlen(trim($htaccess_contents)) === 0) {
+					// empty file, remove it
+					if (!@unlink($uploads_htaccess_file_path)) {
+						throw new wfConfigException($error_message);
+					}
+
+				} elseif (@file_put_contents($uploads_htaccess_file_path, $htaccess_contents, LOCK_EX) === false) {
 					throw new wfConfigException($error_message);
 				}
-				
-			} elseif (@file_put_contents($uploads_htaccess_file_path, $htaccess_contents, LOCK_EX) === false) {
-				throw new wfConfigException($error_message);
 			}
 		}
 		return true;
